Move ModelSelectorList widget to its own file
