+++ moxin-frontend/src/app.rs
@@ -180,10 +180,12 @@ impl LiveRegister for App {
 
         // Chat
         crate::chat::chat_screen::live_design(cx);
-        crate::chat::model_selector::live_design(cx);
         crate::chat::chat_panel::live_design(cx);
         crate::chat::chat_line::live_design(cx);
         crate::chat::chat_line_loading::live_design(cx);
+        crate::chat::model_info::live_design(cx);
+        crate::chat::model_selector::live_design(cx);
+        crate::chat::model_selector_list::live_design(cx);
 
         // My Models
         crate::my_models::my_models_screen::live_design(cx);
