+++ moxin-frontend/src/chat/chat_panel.rs
@@ -1,11 +1,11 @@
 use crate::chat::chat_line::*;
-use crate::chat::model_selector::ModelSelectorAction;
+use model_selector_list::ModelSelectorAction;
 use crate::data::store::Store;
 use crate::my_models::downloaded_files_table::DownloadedFileAction;
 use makepad_widgets::*;
 use moxin_protocol::data::{DownloadedFile, FileID};
 
-use super::model_selector::ModelSelectorWidgetExt;
+use super::{model_selector::ModelSelectorWidgetExt, model_selector_list};
 
 live_design! {
     import makepad_widgets::base::*;
