+++ moxin-frontend/src/chat/mod.rs
@@ -2,4 +2,6 @@ pub mod chat_line;
 pub mod chat_line_loading;
 pub mod chat_panel;
 pub mod chat_screen;
+pub mod model_info;
 pub mod model_selector;
+pub mod model_selector_list;
